+++ libraries/jcb_powers/VDM.Joomla.Gitea/src/Utilities/Http.php
@@ -73,6 +73,5 @@ final class Http extends JoomlaHttp
 
 		$this->setOption('headers', $headers);
 	}
-
 }