+++ libraries/jcb_powers/VDM.Joomla.Gitea/src/Utilities/Uri.php
@@ -135,6 +135,5 @@ final class Uri
 	{
 		$this->version = $version;
 	}
-
 }